+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Notificaiton subscriptions - Shows GiB unit for network in modal
 - Api Page - Persists and loads all filter parameters
+- App - Allow user to choose how to run a docker application
+- App - Rename "Docker compose" accordion to "Compose file"
 
 ## [2.33.3] - 2023-02-27